move setup_mainchain/sidechain to new file
