+++ slk/sidechain.py
@@ -18,191 +18,16 @@ import time
 from multiprocessing import Process, Value
 from typing import Callable, List
 
-from xrpl.models import (
-    AccountSet,
-    AccountSetFlag,
-    Amount,
-    IssuedCurrencyAmount,
-    Memo,
-    Payment,
-    SignerEntry,
-    SignerListSet,
-    TicketCreate,
-    TrustSet,
-)
-from xrpl.utils import xrp_to_drops
+from xrpl.models import Amount, Memo, Payment
 
 import slk.interactive as interactive
 from slk.chain import Chain, configs_for_testnet, single_node_chain
+from slk.chain_setup import setup_mainchain, setup_sidechain
 from slk.common import Account, disable_eprint, eprint
 from slk.config_file import ConfigFile
 from slk.log_analyzer import convert_log
 from slk.sidechain_params import SidechainParams
 from slk.testnet import sidechain_network
-
-MAINCHAIN_DOOR_KEEPER = 0
-SIDECHAIN_DOOR_KEEPER = 1
-UPDATE_SIGNER_LIST = 2
-
-
-def setup_mainchain(
-    mc_chain: Chain, params: SidechainParams, setup_user_accounts: bool = True
-):
-    mc_chain.add_to_keymanager(params.mc_door_account)
-    if setup_user_accounts:
-        mc_chain.add_to_keymanager(params.user_account)
-
-    # mc_chain(LogLevel('fatal'))
-    mc_chain("open")
-
-    # Allow rippling through the genesis account
-    mc_chain(
-        AccountSet(
-            account=params.genesis_account.account_id,
-            set_flag=AccountSetFlag.ASF_DEFAULT_RIPPLE,
-        )
-    )
-    mc_chain.maybe_ledger_accept()
-
-    # Create and fund the mc door account
-    mc_chain(
-        Payment(
-            account=params.genesis_account.account_id,
-            destination=params.mc_door_account.account_id,
-            amount=xrp_to_drops(1_000),
-        )
-    )
-    mc_chain.maybe_ledger_accept()
-
-    # Create a trust line so USD/root account ious can be sent cross chain
-    mc_chain(
-        TrustSet(
-            account=params.mc_door_account.account_id,
-            limit_amount=IssuedCurrencyAmount(
-                value=str(1_000_000),
-                currency="USD",
-                issuer=params.genesis_account.account_id,
-            ),
-        )
-    )
-
-    # set the chain's signer list and disable the master key
-    divide = 4 * len(params.federators)
-    by = 5
-    quorum = (divide + by - 1) // by
-    mc_chain(
-        SignerListSet(
-            account=params.mc_door_account.account_id,
-            signer_quorum=quorum,
-            signer_entries=[
-                SignerEntry(account=federator, signer_weight=1)
-                for federator in params.federators
-            ],
-        )
-    )
-    mc_chain.maybe_ledger_accept()
-    mc_chain(
-        TicketCreate(
-            account=params.mc_door_account.account_id,
-            source_tag=MAINCHAIN_DOOR_KEEPER,
-            ticket_count=1,
-        )
-    )
-    mc_chain.maybe_ledger_accept()
-    mc_chain(
-        TicketCreate(
-            account=params.mc_door_account.account_id,
-            source_tag=SIDECHAIN_DOOR_KEEPER,
-            ticket_count=1,
-        )
-    )
-    mc_chain.maybe_ledger_accept()
-    mc_chain(
-        TicketCreate(
-            account=params.mc_door_account.account_id,
-            source_tag=UPDATE_SIGNER_LIST,
-            ticket_count=1,
-        )
-    )
-    mc_chain.maybe_ledger_accept()
-    mc_chain(
-        AccountSet(
-            account=params.mc_door_account.account_id,
-            set_flag=AccountSetFlag.ASF_DISABLE_MASTER,
-        )
-    )
-    mc_chain.maybe_ledger_accept()
-
-    if setup_user_accounts:
-        # Create and fund a regular user account
-        mc_chain(
-            Payment(
-                account=params.genesis_account.account_id,
-                destination=params.user_account,
-                amount=str(2_000),
-            )
-        )
-        mc_chain.maybe_ledger_accept()
-
-
-def setup_sidechain(
-    sc_chain: Chain, params: SidechainParams, setup_user_accounts: bool = True
-):
-    sc_chain.add_to_keymanager(params.sc_door_account)
-    if setup_user_accounts:
-        sc_chain.add_to_keymanager(params.user_account)
-
-    sc_chain("open")
-
-    # sc_chain(LogLevel('fatal'))
-    # sc_chain(LogLevel('trace', partition='SidechainFederator'))
-
-    # set the chain's signer list and disable the master key
-    divide = 4 * len(params.federators)
-    by = 5
-    quorum = (divide + by - 1) // by
-    sc_chain(
-        SignerListSet(
-            account=params.genesis_account.account_id,
-            signer_quorum=quorum,
-            signer_entries=[
-                SignerEntry(account=federator, signer_weight=1)
-                for federator in params.federators
-            ],
-        )
-    )
-    sc_chain.maybe_ledger_accept()
-    sc_chain(
-        TicketCreate(
-            account=params.genesis_account.account_id,
-            source_tag=MAINCHAIN_DOOR_KEEPER,
-            ticket_count=1,
-        )
-    )
-    sc_chain.maybe_ledger_accept()
-    sc_chain(
-        TicketCreate(
-            account=params.genesis_account.account_id,
-            source_tag=SIDECHAIN_DOOR_KEEPER,
-            ticket_count=1,
-        )
-    )
-    sc_chain.maybe_ledger_accept()
-    sc_chain(
-        TicketCreate(
-            account=params.genesis_account.account_id,
-            source_tag=UPDATE_SIGNER_LIST,
-            ticket_count=1,
-        )
-    )
-    sc_chain.maybe_ledger_accept()
-    sc_chain(
-        AccountSet(
-            account=params.genesis_account.account_id,
-            set_flag=AccountSetFlag.ASF_DISABLE_MASTER,
-        )
-    )
-    sc_chain.maybe_ledger_accept()
 
 
 def _xchain_transfer(
